修正

+++ 2-21/src/Main2.java
@@ -1,9 +1,9 @@
 public class Main2 {
     public static void main(String[] args) {
-        int r = 5;
+        double r = 5;
         calculateCircle(r);
     }
-    static void calculateCircle(int r){
+    static void calculateCircle(double r){
         System.out.printf("円の面積：%.14f%n", Math.PI * r * r);
         System.out.printf("円周：%.14f%n", 2 * Math.PI * r);
     }
